+++ lib/twilio/audiohelper.ts
@@ -164,9 +164,19 @@ class AudioHelper extends EventEmitter {
   private _mediaDevices: AudioHelper.MediaDevicesLike | null;
 
   /**
+   * The microphone permission status
+   */
+  private _microphonePermissionStatus: PermissionStatus | null;
+
+  /**
    * Called with the new input stream when the active input is changed.
    */
   private _onActiveInputChanged: (stream: MediaStream | null) => Promise<void>;
+
+  /**
+   * Handler for microphone permission status change
+   */
+  private _onMicrophonePermissionStatusChanged: () => void;
 
   /**
    * Internal reference to the processed stream
@@ -275,6 +285,20 @@ class AudioHelper extends EventEmitter {
     if (isEnumerationSupported) {
       this._initializeEnumeration();
     }
+
+    // NOTE (kchoy): Currently microphone permissions are not supported in firefox.
+    // https://github.com/mozilla/standards-positions/issues/19#issuecomment-370158947
+    navigator.permissions.query({ name: 'microphone' }).then((microphonePermissionStatus) => {
+      if (microphonePermissionStatus.state !== 'granted') {
+        const handleStateChange = () => {
+          this._updateAvailableDevices();
+          this._stopMicrophonePermissionListener();
+        };
+        microphonePermissionStatus.addEventListener('change', handleStateChange);
+        this._microphonePermissionStatus = microphonePermissionStatus;
+        this._onMicrophonePermissionStatusChanged = handleStateChange;
+      }
+    }).catch((reason) => this._log.warn(`Warning: unable to listen for microphone permission changes. ${reason}`));
   }
 
   /**
@@ -287,6 +311,7 @@ class AudioHelper extends EventEmitter {
     this._destroyProcessedStream();
     this._maybeStopPollingVolume();
     this.removeAllListeners();
+    this._stopMicrophonePermissionListener();
     this._unbind();
   }
 
@@ -797,6 +822,15 @@ class AudioHelper extends EventEmitter {
     return this._inputDevicePromise = setInputDevice().finally(() => {
       this._inputDevicePromise = null;
     });
+  }
+
+  /**
+   * Remove event listener for microphone permissions
+   */
+  private _stopMicrophonePermissionListener(): void {
+    if (this._microphonePermissionStatus?.removeEventListener) {
+      this._microphonePermissionStatus.removeEventListener('change', this._onMicrophonePermissionStatusChanged);
+    }
   }
 
   /**